LibJS: Preserve large substr end indices
Root cause: String.prototype.substr narrowed its end to i32, so a small
suffix beyond INT32_MAX became empty. History shows the spec
rewrite retained a signed length cast from the earlier flat-string code.

Fix and justification: Convert the bounded end to size_t.
UTF-16 lengths fit that type, preserving valid large deferred substrings
without materializing their prefixes.

+++ Libraries/LibJS/Runtime/StringPrototype.cpp
@@ -1572,7 +1572,7 @@ JS_DEFINE_NATIVE_FUNCTION(StringPrototype::substr)
     int_length = clamp(int_length, 0, size);
 
     // 10. Let intEnd be min(intStart + intLength, size).
-    auto int_end = min((i32)(int_start + int_length), size);
+    auto int_end = min(static_cast<size_t>(int_start + int_length), size);
 
     if (int_start >= int_end)
         return PrimitiveString::create(vm, Utf16String {});
